fix(android): SocketProtector.protect 参数类型修正为 Int
gomobile 生成的接口参数为 Int 而非 Long，修正类型匹配。

+++ android/app/src/main/kotlin/cn/mapleafgo/singcast/Mobile.kt
@@ -14,9 +14,9 @@ object Mobile {
     private var vpnService: SingcastVpnService? = null
 
     private val socketProtector = object : SocketProtector {
-        override fun protect(fd: Long): Boolean {
+        override fun protect(fd: Int): Boolean {
             val svc = vpnService ?: return false
-            return svc.protectSocket(fd.toInt())
+            return svc.protectSocket(fd)
         }
     }
